s/store/staticStore/
